requirements.txt
textual==0.47.1
fastapi==0.109.0
uvicorn==0.27.0
pymysql==1.1.0
pymongo==4.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
jinja2==3.1.3
tabulate==0.9.0
python-multipart==0.0.6
pytest==8.0.0
flake8==7.1.1
black==24.8.0
isort==5.13.2